refactor: integrated datetime related
