+++ src/posts/frontmatter.rs
@@ -1,5 +1,4 @@
 use anyhow::{anyhow, Result};
-use chrono::{DateTime, Utc};
 use itertools::Itertools;
 use linked_hash_map::LinkedHashMap;
 use uuid::Uuid;
@@ -7,47 +6,10 @@ use yaml_rust::{Yaml, YamlLoader};
 
 use crate::{
     posts::Lang,
-    text_engine::{
-        datetime::{parse_datetime, DateTimeFormat},
-        schema::PostField,
-    },
+    text_engine::schema::PostField,
+    datetime::{parse_datetime, DateTimeWithFormat},
 };
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct DateTimeWithFormat {
-    datetime: DateTime<Utc>,
-    format: DateTimeFormat,
-}
-
-impl DateTimeWithFormat {
-    pub fn new(datetime: DateTime<Utc>, format: DateTimeFormat) -> Self {
-        Self { datetime, format }
-    }
-
-    pub fn now(format: &DateTimeFormat) -> Self {
-        Self::new(Utc::now(), format.to_owned())
-    }
-
-    pub fn datetime(&self) -> DateTime<Utc> {
-        self.datetime
-    }
-
-    pub fn format(&self) -> DateTimeFormat {
-        self.format.clone()
-    }
-}
-
-impl ToString for DateTimeWithFormat {
-    fn to_string(&self) -> String {
-        self.format.format(self.datetime)
-    }
-}
-
-impl Default for DateTimeWithFormat {
-    fn default() -> Self {
-        Self::now(&DateTimeFormat::RFC3339)
-    }
-}
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct FrontMatter {
@@ -228,14 +190,14 @@ fn get_str_from_yaml(doc: &Yaml, field: PostField) -> Result<String> {
 
 pub fn parse_date_with_format(date_string: &str) -> DateTimeWithFormat {
     match parse_datetime(date_string, None) {
-        Ok((format, datetime)) => DateTimeWithFormat { datetime, format },
+        Ok((format, datetime)) => DateTimeWithFormat::new(datetime, format),
         Err(e) => panic!("{}", e),
     }
 }
 
 pub fn parse_date_from_yaml(doc: &Yaml, key: &str) -> Option<DateTimeWithFormat> {
     doc[key].as_str().map(|s| match parse_datetime(s, None) {
-        Ok((format, datetime)) => DateTimeWithFormat { datetime, format },
+        Ok((format, datetime)) => DateTimeWithFormat::new(datetime, format),
         Err(e) => panic!("{}", e),
     })
 }
@@ -368,7 +330,10 @@ pub fn split_frontmatter_and_content(text: &str) -> (Option<FrontMatter>, &str) 
 
 #[cfg(test)]
 mod test {
+    use chrono::Utc;
     use yaml_rust::YamlEmitter;
+
+    use crate::datetime::DateTimeFormat;
 
     use super::*;
     #[test]
