+++ src/server/route/posts.rs
@@ -13,7 +13,7 @@ use crate::text_engine::{
     schema::{FieldGetter, PostField},
 };
 
-use crate::text_engine::query::OrderBy;
+use crate::datetime;
 
 #[get("/post/uuid/{uuid}")]
 async fn get_post_by_id(index: web::Data<Index>, uuid: web::Path<String>) -> HttpResponse {
@@ -70,7 +70,7 @@ pub struct GetPostsQueryParams {
     lang: Option<String>,
     category: Option<String>,
     tag: Option<String>,
-    order_by: Option<OrderBy>,
+    order_by: Option<datetime::OrderBy>,
     order: Option<Order>,
 }
 
@@ -95,7 +95,7 @@ impl GetPostsQueryParams {
         .collect()
     }
 
-    pub fn order_by(&self) -> Option<OrderBy> {
+    pub fn order_by(&self) -> Option<datetime::OrderBy> {
         self.order_by.to_owned()
     }
