+++ src/text_engine/mod.rs
@@ -1,4 +1,3 @@
-pub mod datetime;
 pub mod index;
 pub mod query;
 pub mod schema;
